+++ test/fuzz/RngLockRotationDeterminism.t.sol
@@ -247,7 +247,7 @@ contract RngLockRotationDeterminism is DeployProtocol {
     ///      identical digest whether or not a rotation occurred mid-window.
     function _captureDailyVrfDigest() internal view returns (bytes32) {
         uint32 today = game.currentDayView();
-        uint256 dayWord = game.rngWordForDay(today);
+        uint256 dayWord = game.rngWordForDay(uint24(today));
         uint256 rngWordCurrent = _readRngWordCurrent();
         return keccak256(abi.encode(today, dayWord, rngWordCurrent));
     }
